+++ gen/com/intellij/idea/plugin/hybris/impex/ImpexParser.java
@@ -1,4 +1,8 @@
 /*
+ * ----------------------------------------------------------------
+ * --- WARNING: THIS FILE IS GENERATED AND WILL BE OVERWRITTEN! ---
+ * ----------------------------------------------------------------
+ *
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
  * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
  *
@@ -1062,70 +1066,143 @@ public class ImpexParser implements PsiParser, LightPsiParser {
   }
 
   /* ********************************************************** */
-  // user_rights_start CRLF (user_rights_header_line CRLF)? (user_rights_value_line CRLF)* user_rights_end
+  // user_rights_start (PARAMETERS_SEPARATOR)* CRLF+ (user_rights_header_line CRLF+) (user_rights_value_line CRLF+)* user_rights_end
   public static boolean user_rights(PsiBuilder b, int l) {
     if (!recursion_guard_(b, l, "user_rights")) return false;
     boolean r, p;
     Marker m = enter_section_(b, l, _NONE_, USER_RIGHTS, "<user rights>");
     r = user_rights_start(b, l + 1);
     p = r; // pin = 1
-    r = r && report_error_(b, consumeToken(b, CRLF));
+    r = r && report_error_(b, user_rights_1(b, l + 1));
     r = p && report_error_(b, user_rights_2(b, l + 1)) && r;
     r = p && report_error_(b, user_rights_3(b, l + 1)) && r;
+    r = p && report_error_(b, user_rights_4(b, l + 1)) && r;
     r = p && user_rights_end(b, l + 1) && r;
     exit_section_(b, l, m, r, p, ImpexParser::recover_root);
     return r || p;
   }
 
-  // (user_rights_header_line CRLF)?
-  private static boolean user_rights_2(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "user_rights_2")) return false;
-    user_rights_2_0(b, l + 1);
-    return true;
-  }
-
-  // user_rights_header_line CRLF
-  private static boolean user_rights_2_0(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "user_rights_2_0")) return false;
-    boolean r;
-    Marker m = enter_section_(b);
-    r = user_rights_header_line(b, l + 1);
-    r = r && consumeToken(b, CRLF);
-    exit_section_(b, m, null, r);
-    return r;
-  }
-
-  // (user_rights_value_line CRLF)*
-  private static boolean user_rights_3(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "user_rights_3")) return false;
+  // (PARAMETERS_SEPARATOR)*
+  private static boolean user_rights_1(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_1")) return false;
     while (true) {
       int c = current_position_(b);
-      if (!user_rights_3_0(b, l + 1)) break;
-      if (!empty_element_parsed_guard_(b, "user_rights_3", c)) break;
+      if (!consumeToken(b, PARAMETERS_SEPARATOR)) break;
+      if (!empty_element_parsed_guard_(b, "user_rights_1", c)) break;
     }
     return true;
   }
 
-  // user_rights_value_line CRLF
-  private static boolean user_rights_3_0(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "user_rights_3_0")) return false;
+  // CRLF+
+  private static boolean user_rights_2(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_2")) return false;
+    boolean r;
+    Marker m = enter_section_(b);
+    r = consumeToken(b, CRLF);
+    while (r) {
+      int c = current_position_(b);
+      if (!consumeToken(b, CRLF)) break;
+      if (!empty_element_parsed_guard_(b, "user_rights_2", c)) break;
+    }
+    exit_section_(b, m, null, r);
+    return r;
+  }
+
+  // user_rights_header_line CRLF+
+  private static boolean user_rights_3(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_3")) return false;
+    boolean r;
+    Marker m = enter_section_(b);
+    r = user_rights_header_line(b, l + 1);
+    r = r && user_rights_3_1(b, l + 1);
+    exit_section_(b, m, null, r);
+    return r;
+  }
+
+  // CRLF+
+  private static boolean user_rights_3_1(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_3_1")) return false;
+    boolean r;
+    Marker m = enter_section_(b);
+    r = consumeToken(b, CRLF);
+    while (r) {
+      int c = current_position_(b);
+      if (!consumeToken(b, CRLF)) break;
+      if (!empty_element_parsed_guard_(b, "user_rights_3_1", c)) break;
+    }
+    exit_section_(b, m, null, r);
+    return r;
+  }
+
+  // (user_rights_value_line CRLF+)*
+  private static boolean user_rights_4(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_4")) return false;
+    while (true) {
+      int c = current_position_(b);
+      if (!user_rights_4_0(b, l + 1)) break;
+      if (!empty_element_parsed_guard_(b, "user_rights_4", c)) break;
+    }
+    return true;
+  }
+
+  // user_rights_value_line CRLF+
+  private static boolean user_rights_4_0(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_4_0")) return false;
     boolean r;
     Marker m = enter_section_(b);
     r = user_rights_value_line(b, l + 1);
-    r = r && consumeToken(b, CRLF);
+    r = r && user_rights_4_0_1(b, l + 1);
+    exit_section_(b, m, null, r);
+    return r;
+  }
+
+  // CRLF+
+  private static boolean user_rights_4_0_1(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_4_0_1")) return false;
+    boolean r;
+    Marker m = enter_section_(b);
+    r = consumeToken(b, CRLF);
+    while (r) {
+      int c = current_position_(b);
+      if (!consumeToken(b, CRLF)) break;
+      if (!empty_element_parsed_guard_(b, "user_rights_4_0_1", c)) break;
+    }
     exit_section_(b, m, null, r);
     return r;
   }
 
   /* ********************************************************** */
-  // END_USERRIGHTS
+  // END_USERRIGHTS (PARAMETERS_SEPARATOR)*
   public static boolean user_rights_end(PsiBuilder b, int l) {
     if (!recursion_guard_(b, l, "user_rights_end")) return false;
     if (!nextTokenIs(b, END_USERRIGHTS)) return false;
     boolean r;
     Marker m = enter_section_(b);
     r = consumeToken(b, END_USERRIGHTS);
+    r = r && user_rights_end_1(b, l + 1);
     exit_section_(b, m, USER_RIGHTS_END, r);
+    return r;
+  }
+
+  // (PARAMETERS_SEPARATOR)*
+  private static boolean user_rights_end_1(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_end_1")) return false;
+    while (true) {
+      int c = current_position_(b);
+      if (!consumeToken(b, PARAMETERS_SEPARATOR)) break;
+      if (!empty_element_parsed_guard_(b, "user_rights_end_1", c)) break;
+    }
+    return true;
+  }
+
+  /* ********************************************************** */
+  // user_rights_value
+  public static boolean user_rights_first_value_group(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_first_value_group")) return false;
+    boolean r;
+    Marker m = enter_section_(b, l, _NONE_, USER_RIGHTS_FIRST_VALUE_GROUP, "<user rights first value group>");
+    r = user_rights_value(b, l + 1);
+    exit_section_(b, l, m, r, false, ImpexParser::user_rights_value_group_recover);
     return r;
   }
 
@@ -1138,7 +1215,7 @@ public class ImpexParser implements PsiParser, LightPsiParser {
     r = user_rights_header_parameter(b, l + 1);
     p = r; // pin = 1
     r = r && user_rights_header_line_1(b, l + 1);
-    exit_section_(b, l, m, r, p, ImpexParser::user_rights_header_line_recover);
+    exit_section_(b, l, m, r, p, ImpexParser::not_line_break);
     return r || p;
   }
 
@@ -1165,35 +1242,32 @@ public class ImpexParser implements PsiParser, LightPsiParser {
   }
 
   /* ********************************************************** */
-  // !(CRLF | user_rights_end | user_rights_value_line )
-  static boolean user_rights_header_line_recover(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "user_rights_header_line_recover")) return false;
-    boolean r;
-    Marker m = enter_section_(b, l, _NOT_);
-    r = !user_rights_header_line_recover_0(b, l + 1);
-    exit_section_(b, l, m, r, false, null);
-    return r;
-  }
-
-  // CRLF | user_rights_end | user_rights_value_line
-  private static boolean user_rights_header_line_recover_0(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "user_rights_header_line_recover_0")) return false;
-    boolean r;
-    Marker m = enter_section_(b);
-    r = consumeToken(b, CRLF);
-    if (!r) r = user_rights_end(b, l + 1);
-    if (!r) r = user_rights_value_line(b, l + 1);
-    exit_section_(b, m, null, r);
-    return r;
-  }
-
-  /* ********************************************************** */
-  // USER_RIGHTS_HEADER_PARAMETER_NAME
+  // TYPE
+  //  | UID
+  //  | MEMBEROFGROUPS
+  //  | PASSWORD
+  //  | TARGET
+  //  | READ
+  //  | CHANGE
+  //  | CREATE
+  //  | DELETE
+  //  | REMOVE
+  //  | CHANGE_PERM
   public static boolean user_rights_header_parameter(PsiBuilder b, int l) {
     if (!recursion_guard_(b, l, "user_rights_header_parameter")) return false;
     boolean r;
     Marker m = enter_section_(b, l, _NONE_, USER_RIGHTS_HEADER_PARAMETER, "<user rights header parameter>");
-    r = consumeToken(b, USER_RIGHTS_HEADER_PARAMETER_NAME);
+    r = consumeToken(b, TYPE);
+    if (!r) r = consumeToken(b, UID);
+    if (!r) r = consumeToken(b, MEMBEROFGROUPS);
+    if (!r) r = consumeToken(b, PASSWORD);
+    if (!r) r = consumeToken(b, TARGET);
+    if (!r) r = consumeToken(b, READ);
+    if (!r) r = consumeToken(b, CHANGE);
+    if (!r) r = consumeToken(b, CREATE);
+    if (!r) r = consumeToken(b, DELETE);
+    if (!r) r = consumeToken(b, REMOVE);
+    if (!r) r = consumeToken(b, CHANGE_PERM);
     exit_section_(b, l, m, r, false, ImpexParser::not_line_break_or_parameters_separator);
     return r;
   }
@@ -1211,13 +1285,13 @@ public class ImpexParser implements PsiParser, LightPsiParser {
   }
 
   /* ********************************************************** */
-  // USER_RIGHTS_VALUE DOT USER_RIGHTS_VALUE
+  // FIELD_VALUE DOT FIELD_VALUE
   static boolean user_rights_type_attribute(PsiBuilder b, int l) {
     if (!recursion_guard_(b, l, "user_rights_type_attribute")) return false;
-    if (!nextTokenIs(b, USER_RIGHTS_VALUE)) return false;
+    if (!nextTokenIs(b, FIELD_VALUE)) return false;
     boolean r, p;
     Marker m = enter_section_(b, l, _NONE_);
-    r = consumeTokens(b, 2, USER_RIGHTS_VALUE, DOT, USER_RIGHTS_VALUE);
+    r = consumeTokens(b, 2, FIELD_VALUE, DOT, FIELD_VALUE);
     p = r; // pin = 2
     exit_section_(b, l, m, r, p, null);
     return r || p;
@@ -1227,116 +1301,142 @@ public class ImpexParser implements PsiParser, LightPsiParser {
   // PERMISSION_DENIED
   //     | PERMISSION_ALLOWED
   //     | user_rights_type_attribute
-  //     | USER_RIGHTS_VALUE
-  static boolean user_rights_value_dec(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "user_rights_value_dec")) return false;
+  //     | FIELD_VALUE (COMMA FIELD_VALUE)*
+  public static boolean user_rights_value(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_value")) return false;
     boolean r;
+    Marker m = enter_section_(b, l, _NONE_, USER_RIGHTS_VALUE, "<user rights value>");
     r = consumeToken(b, PERMISSION_DENIED);
     if (!r) r = consumeToken(b, PERMISSION_ALLOWED);
     if (!r) r = user_rights_type_attribute(b, l + 1);
-    if (!r) r = consumeToken(b, USER_RIGHTS_VALUE);
+    if (!r) r = user_rights_value_3(b, l + 1);
+    exit_section_(b, l, m, r, false, ImpexParser::user_rights_value_dec_recover);
     return r;
   }
 
-  /* ********************************************************** */
-  // user_rights_value_dec (COMMA user_rights_value_dec)*
-  public static boolean user_rights_value_group(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "user_rights_value_group")) return false;
+  // FIELD_VALUE (COMMA FIELD_VALUE)*
+  private static boolean user_rights_value_3(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_value_3")) return false;
     boolean r;
-    Marker m = enter_section_(b, l, _NONE_, USER_RIGHTS_VALUE_GROUP, "<user rights value group>");
-    r = user_rights_value_dec(b, l + 1);
-    r = r && user_rights_value_group_1(b, l + 1);
-    exit_section_(b, l, m, r, false, null);
+    Marker m = enter_section_(b);
+    r = consumeToken(b, FIELD_VALUE);
+    r = r && user_rights_value_3_1(b, l + 1);
+    exit_section_(b, m, null, r);
     return r;
   }
 
-  // (COMMA user_rights_value_dec)*
-  private static boolean user_rights_value_group_1(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "user_rights_value_group_1")) return false;
+  // (COMMA FIELD_VALUE)*
+  private static boolean user_rights_value_3_1(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_value_3_1")) return false;
     while (true) {
       int c = current_position_(b);
-      if (!user_rights_value_group_1_0(b, l + 1)) break;
-      if (!empty_element_parsed_guard_(b, "user_rights_value_group_1", c)) break;
+      if (!user_rights_value_3_1_0(b, l + 1)) break;
+      if (!empty_element_parsed_guard_(b, "user_rights_value_3_1", c)) break;
     }
     return true;
   }
 
-  // COMMA user_rights_value_dec
-  private static boolean user_rights_value_group_1_0(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "user_rights_value_group_1_0")) return false;
+  // COMMA FIELD_VALUE
+  private static boolean user_rights_value_3_1_0(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_value_3_1_0")) return false;
     boolean r;
     Marker m = enter_section_(b);
-    r = consumeToken(b, COMMA);
-    r = r && user_rights_value_dec(b, l + 1);
+    r = consumeTokens(b, 0, COMMA, FIELD_VALUE);
     exit_section_(b, m, null, r);
     return r;
   }
 
   /* ********************************************************** */
-  // user_rights_value_group? (FIELD_VALUE_SEPARATOR user_rights_value_group?)*
+  // !(CRLF | FIELD_VALUE_SEPARATOR)
+  static boolean user_rights_value_dec_recover(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_value_dec_recover")) return false;
+    boolean r;
+    Marker m = enter_section_(b, l, _NOT_);
+    r = !user_rights_value_dec_recover_0(b, l + 1);
+    exit_section_(b, l, m, r, false, null);
+    return r;
+  }
+
+  // CRLF | FIELD_VALUE_SEPARATOR
+  private static boolean user_rights_value_dec_recover_0(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_value_dec_recover_0")) return false;
+    boolean r;
+    r = consumeToken(b, CRLF);
+    if (!r) r = consumeToken(b, FIELD_VALUE_SEPARATOR);
+    return r;
+  }
+
+  /* ********************************************************** */
+  // FIELD_VALUE_SEPARATOR user_rights_value?
+  public static boolean user_rights_value_group(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_value_group")) return false;
+    boolean r, p;
+    Marker m = enter_section_(b, l, _NONE_, USER_RIGHTS_VALUE_GROUP, "<user rights value group>");
+    r = consumeToken(b, FIELD_VALUE_SEPARATOR);
+    p = r; // pin = 1
+    r = r && user_rights_value_group_1(b, l + 1);
+    exit_section_(b, l, m, r, p, ImpexParser::user_rights_value_group_recover);
+    return r || p;
+  }
+
+  // user_rights_value?
+  private static boolean user_rights_value_group_1(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_value_group_1")) return false;
+    user_rights_value(b, l + 1);
+    return true;
+  }
+
+  /* ********************************************************** */
+  // !(CRLF | FIELD_VALUE_SEPARATOR)
+  static boolean user_rights_value_group_recover(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_value_group_recover")) return false;
+    boolean r;
+    Marker m = enter_section_(b, l, _NOT_);
+    r = !user_rights_value_group_recover_0(b, l + 1);
+    exit_section_(b, l, m, r, false, null);
+    return r;
+  }
+
+  // CRLF | FIELD_VALUE_SEPARATOR
+  private static boolean user_rights_value_group_recover_0(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "user_rights_value_group_recover_0")) return false;
+    boolean r;
+    r = consumeToken(b, CRLF);
+    if (!r) r = consumeToken(b, FIELD_VALUE_SEPARATOR);
+    return r;
+  }
+
+  /* ********************************************************** */
+  // user_rights_first_value_group? user_rights_value_group+
   public static boolean user_rights_value_line(PsiBuilder b, int l) {
     if (!recursion_guard_(b, l, "user_rights_value_line")) return false;
     boolean r;
     Marker m = enter_section_(b, l, _NONE_, USER_RIGHTS_VALUE_LINE, "<user rights value line>");
     r = user_rights_value_line_0(b, l + 1);
     r = r && user_rights_value_line_1(b, l + 1);
-    exit_section_(b, l, m, r, false, ImpexParser::user_rights_value_line_recover);
+    exit_section_(b, l, m, r, false, ImpexParser::not_line_break);
     return r;
   }
 
-  // user_rights_value_group?
+  // user_rights_first_value_group?
   private static boolean user_rights_value_line_0(PsiBuilder b, int l) {
     if (!recursion_guard_(b, l, "user_rights_value_line_0")) return false;
-    user_rights_value_group(b, l + 1);
+    user_rights_first_value_group(b, l + 1);
     return true;
   }
 
-  // (FIELD_VALUE_SEPARATOR user_rights_value_group?)*
+  // user_rights_value_group+
   private static boolean user_rights_value_line_1(PsiBuilder b, int l) {
     if (!recursion_guard_(b, l, "user_rights_value_line_1")) return false;
-    while (true) {
-      int c = current_position_(b);
-      if (!user_rights_value_line_1_0(b, l + 1)) break;
-      if (!empty_element_parsed_guard_(b, "user_rights_value_line_1", c)) break;
-    }
-    return true;
-  }
-
-  // FIELD_VALUE_SEPARATOR user_rights_value_group?
-  private static boolean user_rights_value_line_1_0(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "user_rights_value_line_1_0")) return false;
     boolean r;
     Marker m = enter_section_(b);
-    r = consumeToken(b, FIELD_VALUE_SEPARATOR);
-    r = r && user_rights_value_line_1_0_1(b, l + 1);
+    r = user_rights_value_group(b, l + 1);
+    while (r) {
+      int c = current_position_(b);
+      if (!user_rights_value_group(b, l + 1)) break;
+      if (!empty_element_parsed_guard_(b, "user_rights_value_line_1", c)) break;
+    }
     exit_section_(b, m, null, r);
-    return r;
-  }
-
-  // user_rights_value_group?
-  private static boolean user_rights_value_line_1_0_1(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "user_rights_value_line_1_0_1")) return false;
-    user_rights_value_group(b, l + 1);
-    return true;
-  }
-
-  /* ********************************************************** */
-  // !(CRLF | user_rights_end)
-  static boolean user_rights_value_line_recover(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "user_rights_value_line_recover")) return false;
-    boolean r;
-    Marker m = enter_section_(b, l, _NOT_);
-    r = !user_rights_value_line_recover_0(b, l + 1);
-    exit_section_(b, l, m, r, false, null);
-    return r;
-  }
-
-  // CRLF | user_rights_end
-  private static boolean user_rights_value_line_recover_0(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "user_rights_value_line_recover_0")) return false;
-    boolean r;
-    r = consumeToken(b, CRLF);
-    if (!r) r = user_rights_end(b, l + 1);
     return r;
   }