+++ gen/com/intellij/idea/plugin/hybris/impex/psi/ImpexUserRightsFirstValueGroup.java
@@ -25,18 +25,12 @@ import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
 
-public interface ImpexUserRightsValueGroup extends ImpexUserRightsAwarePsiElement {
+public interface ImpexUserRightsFirstValueGroup extends ImpexUserRightsAwarePsiElement {
 
-  @Nullable
+  @NotNull
   ImpexUserRightsValue getUserRightsValue();
 
   @Nullable
-  ImpexUserRightsValueLine getValueLine();
-
-  @Nullable
-  Integer getColumnNumber();
-
-  @Nullable
-  ImpexUserRightsHeaderParameter getHeaderParameter();
+  ImpexValueLine getValueLine();
 
 }