+++ gen/com/intellij/idea/plugin/hybris/impex/psi/ImpexUserRightsHeaderLine.java
@@ -25,9 +25,12 @@ import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
 
-public interface ImpexUserRightsHeaderLine extends PsiElement {
+public interface ImpexUserRightsHeaderLine extends ImpexUserRightsAwarePsiElement {
 
   @NotNull
   List<ImpexUserRightsHeaderParameter> getUserRightsHeaderParameterList();
+
+  @Nullable
+  ImpexUserRightsHeaderParameter getHeaderParameter(int index);
 
 }