+++ gen/com/intellij/idea/plugin/hybris/impex/psi/ImpexUserRightsStart.java
@@ -25,6 +25,6 @@ import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
 
-public interface ImpexUserRightsStart extends PsiElement {
+public interface ImpexUserRightsStart extends ImpexUserRightsAwarePsiElement {
 
 }