+++ gen/com/intellij/idea/plugin/hybris/impex/psi/ImpexVisitor.java
@@ -120,27 +120,35 @@ public class ImpexVisitor extends PsiElementVisitor {
   }
 
   public void visitUserRightsEnd(@NotNull ImpexUserRightsEnd o) {
-    visitPsiElement(o);
+    visitUserRightsAwarePsiElement(o);
+  }
+
+  public void visitUserRightsFirstValueGroup(@NotNull ImpexUserRightsFirstValueGroup o) {
+    visitUserRightsAwarePsiElement(o);
   }
 
   public void visitUserRightsHeaderLine(@NotNull ImpexUserRightsHeaderLine o) {
-    visitPsiElement(o);
+    visitUserRightsAwarePsiElement(o);
   }
 
   public void visitUserRightsHeaderParameter(@NotNull ImpexUserRightsHeaderParameter o) {
-    visitPsiElement(o);
+    visitUserRightsAwarePsiElement(o);
   }
 
   public void visitUserRightsStart(@NotNull ImpexUserRightsStart o) {
-    visitPsiElement(o);
+    visitUserRightsAwarePsiElement(o);
+  }
+
+  public void visitUserRightsValue(@NotNull ImpexUserRightsValue o) {
+    visitUserRightsAwarePsiElement(o);
   }
 
   public void visitUserRightsValueGroup(@NotNull ImpexUserRightsValueGroup o) {
-    visitPsiElement(o);
+    visitUserRightsAwarePsiElement(o);
   }
 
   public void visitUserRightsValueLine(@NotNull ImpexUserRightsValueLine o) {
-    visitPsiElement(o);
+    visitUserRightsAwarePsiElement(o);
   }
 
   public void visitValue(@NotNull ImpexValue o) {
@@ -156,6 +164,10 @@ public class ImpexVisitor extends PsiElementVisitor {
   }
 
   public void visitPsiNamedElement(@NotNull ImpexPsiNamedElement o) {
+    visitPsiElement(o);
+  }
+
+  public void visitUserRightsAwarePsiElement(@NotNull ImpexUserRightsAwarePsiElement o) {
     visitPsiElement(o);
   }