+++ gen/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexUserRightsEndImpl.java
@@ -28,10 +28,9 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
 import static com.intellij.idea.plugin.hybris.impex.psi.ImpexTypes.*;
-import com.intellij.extapi.psi.ASTWrapperPsiElement;
 import com.intellij.idea.plugin.hybris.impex.psi.*;
 
-public class ImpexUserRightsEndImpl extends ASTWrapperPsiElement implements ImpexUserRightsEnd {
+public class ImpexUserRightsEndImpl extends ImpexUserRightsAwarePsiElementImpl implements ImpexUserRightsEnd {
 
   public ImpexUserRightsEndImpl(@NotNull ASTNode node) {
     super(node);