+++ gen/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexUserRightsHeaderParameterImpl.java
@@ -28,10 +28,10 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
 import static com.intellij.idea.plugin.hybris.impex.psi.ImpexTypes.*;
-import com.intellij.extapi.psi.ASTWrapperPsiElement;
 import com.intellij.idea.plugin.hybris.impex.psi.*;
+import java.util.Collection;
 
-public class ImpexUserRightsHeaderParameterImpl extends ASTWrapperPsiElement implements ImpexUserRightsHeaderParameter {
+public class ImpexUserRightsHeaderParameterImpl extends ImpexUserRightsAwarePsiElementImpl implements ImpexUserRightsHeaderParameter {
 
   public ImpexUserRightsHeaderParameterImpl(@NotNull ASTNode node) {
     super(node);
@@ -45,6 +45,24 @@ public class ImpexUserRightsHeaderParameterImpl extends ASTWrapperPsiElement imp
   public void accept(@NotNull PsiElementVisitor visitor) {
     if (visitor instanceof ImpexVisitor) accept((ImpexVisitor)visitor);
     else super.accept(visitor);
+  }
+
+  @Override
+  @Nullable
+  public ImpexUserRightsHeaderLine getHeaderLine() {
+    return ImpexPsiUtil.getHeaderLine(this);
+  }
+
+  @Override
+  @Nullable
+  public Integer getColumnNumber() {
+    return ImpexPsiUtil.getColumnNumber(this);
+  }
+
+  @Override
+  @NotNull
+  public Collection<ImpexUserRightsValueGroup> getValueGroups() {
+    return ImpexPsiUtil.getValueGroups(this);
   }
 
 }