+++ gen/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexUserRightsImpl.java
@@ -30,6 +30,7 @@ import com.intellij.psi.util.PsiTreeUtil;
 import static com.intellij.idea.plugin.hybris.impex.psi.ImpexTypes.*;
 import com.intellij.extapi.psi.ASTWrapperPsiElement;
 import com.intellij.idea.plugin.hybris.impex.psi.*;
+import java.util.Collection;
 
 public class ImpexUserRightsImpl extends ASTWrapperPsiElement implements ImpexUserRights {
 
@@ -69,6 +70,12 @@ public class ImpexUserRightsImpl extends ASTWrapperPsiElement implements ImpexUs
   @NotNull
   public List<ImpexUserRightsValueLine> getUserRightsValueLineList() {
     return PsiTreeUtil.getChildrenOfTypeAsList(this, ImpexUserRightsValueLine.class);
+  }
+
+  @Override
+  @NotNull
+  public Collection<ImpexUserRightsValueGroup> getValueGroups(int index) {
+    return ImpexPsiUtil.getValueGroups(this, index);
   }
 
 }