+++ gen/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexUserRightsValueGroupImpl.java
@@ -28,10 +28,9 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
 import static com.intellij.idea.plugin.hybris.impex.psi.ImpexTypes.*;
-import com.intellij.extapi.psi.ASTWrapperPsiElement;
 import com.intellij.idea.plugin.hybris.impex.psi.*;
 
-public class ImpexUserRightsValueGroupImpl extends ASTWrapperPsiElement implements ImpexUserRightsValueGroup {
+public class ImpexUserRightsValueGroupImpl extends ImpexUserRightsAwarePsiElementImpl implements ImpexUserRightsValueGroup {
 
   public ImpexUserRightsValueGroupImpl(@NotNull ASTNode node) {
     super(node);
@@ -45,6 +44,30 @@ public class ImpexUserRightsValueGroupImpl extends ASTWrapperPsiElement implemen
   public void accept(@NotNull PsiElementVisitor visitor) {
     if (visitor instanceof ImpexVisitor) accept((ImpexVisitor)visitor);
     else super.accept(visitor);
+  }
+
+  @Override
+  @Nullable
+  public ImpexUserRightsValue getUserRightsValue() {
+    return findChildByClass(ImpexUserRightsValue.class);
+  }
+
+  @Override
+  @Nullable
+  public ImpexUserRightsValueLine getValueLine() {
+    return ImpexPsiUtil.getValueLine(this);
+  }
+
+  @Override
+  @Nullable
+  public Integer getColumnNumber() {
+    return ImpexPsiUtil.getColumnNumber(this);
+  }
+
+  @Override
+  @Nullable
+  public ImpexUserRightsHeaderParameter getHeaderParameter() {
+    return ImpexPsiUtil.getHeaderParameter(this);
   }
 
 }