+++ src/com/intellij/idea/plugin/hybris/impex/psi/ImpexPsiUtil.kt
@@ -29,6 +29,7 @@ import com.intellij.openapi.util.text.StringUtil
 import com.intellij.psi.PsiElement
 import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.psi.util.childrenOfType
+import com.intellij.psi.util.parentOfType
 import com.intellij.psi.util.siblings
 
 fun getHeaderLine(element: ImpexFullHeaderParameter): ImpexHeaderLine? = PsiTreeUtil
@@ -116,4 +117,77 @@ fun getConfigPropertyKey(element: ImpexMacroUsageDec): String? {
         .findMacroProperty(element.project, propertyKey)
         ?.key
         ?: element.text.replace(HybrisConstants.IMPEX_CONFIG_COMPLETE_PREFIX, "")
+}
+
+// ------------------------------------------
+//              User Rights
+// ------------------------------------------
+fun getValueGroups(element: ImpexUserRights, index: Int): Collection<ImpexUserRightsValueGroup> = element
+    .userRightsValueLineList
+    .mapNotNull { it.getValueGroup(index) }
+
+fun getValueGroup(element: ImpexUserRightsValueLine, index: Int): ImpexUserRightsValueGroup? = element
+    .userRightsValueGroupList
+    .getOrNull(index)
+
+fun getHeaderParameter(element: ImpexUserRightsHeaderLine, index: Int): ImpexUserRightsHeaderParameter? = element
+    .userRightsHeaderParameterList
+    .getOrNull(index)
+
+fun getValueLine(element: ImpexUserRightsValueGroup): ImpexUserRightsValueLine? = element
+    .parentOfType<ImpexUserRightsValueLine>()
+
+fun getColumnNumber(element: ImpexUserRightsValueGroup): Int? = element
+    .valueLine
+    ?.let { valueLine ->
+        valueLine.userRightsValueGroupList.indexOf(element)
+            .takeIf { it != -1 }
+            ?.let {
+                // we always have to plus one column, because first value group is not part of the list
+                it + 1
+            }
+    }
+
+fun getHeaderParameter(element: ImpexUserRightsValueGroup): ImpexUserRightsHeaderParameter? = element
+    .columnNumber
+    ?.let {
+        element.getUserRights()
+            ?.userRightsHeaderLine
+            ?.getHeaderParameter(it)
+    }
+
+fun getHeaderParameter(element: ImpexUserRightsValue): ImpexUserRightsHeaderParameter? = when (val parent = element.parent) {
+    is ImpexUserRightsFirstValueGroup -> {
+        parent.getUserRights()
+            ?.userRightsHeaderLine
+            ?.getHeaderParameter(0)
+    }
+
+    is ImpexUserRightsValueGroup -> {
+        parent
+            .columnNumber
+            ?.let {
+                element.getUserRights()
+                    ?.userRightsHeaderLine
+                    ?.getHeaderParameter(it)
+
+            }
+    }
+
+    else -> null
+}
+
+fun getHeaderLine(element: ImpexUserRightsHeaderParameter): ImpexUserRightsHeaderLine? = element
+    .parentOfType<ImpexUserRightsHeaderLine>()
+
+fun getColumnNumber(element: ImpexUserRightsHeaderParameter): Int? = element
+    .headerLine
+    ?.userRightsHeaderParameterList
+    ?.indexOf(element)
+    ?.takeIf { it != -1 }
+
+fun getValueGroups(element: ImpexUserRightsHeaderParameter): Collection<ImpexUserRightsValueGroup> {
+    val columnNumber = element.columnNumber ?: return emptyList()
+    val userRights = element.getUserRights() ?: return emptyList()
+    return userRights.getValueGroups(columnNumber)
 }